Enhance layout for team members 5 horizontal
Added a new media query for larger screens. Increased the width of the
container for the team section, decreased the width of the team section
list items, and removed the clear property from the 4 team section list
item.

I also added these to the custom css on the site. Adding it here to have
it in a checkin and in the repo for documentation.